fix(infrastructure): :bug: auth fix
