+++ apps/api/src/auth/infrastructure/controllers/auth.controller.ts
@@ -3,52 +3,33 @@ import {
   Post,
   Body,
   Inject,
-  Get,
-  UnauthorizedException,
   Put,
 } from '@nestjs/common';
 import { ApiResponse, ApiTags } from '@nestjs/swagger';
-import { Model } from 'mongoose';
 import {
   CheckCodeDto,
-  LoginDto,
   RequestVerificationCodeDto,
-  SignUpDto,
   ResetPasswordDto,
 } from './dtos';
 import {
   AUTH_REPOSITORY,
   CODE_GENERATOR,
-  JWT_SERVICE,
   VERIFICATION_EMAIL_HANDLER,
 } from '../constants';
 import {
   CheckVerificationCodeCommand,
   CredentialsRepository,
-  LoginCommand,
   RequestVerificationCodeCommand,
   ResetPasswordCommand,
-  SignUpCommand,
 } from '../../application';
 import {
   BCRYPT_SERVICE,
   CryptoService,
-  UUIDGENERATOR,
-  IdGenerator,
   EmailHandler,
   CodeGenerator,
   SuccessBasedResponse,
-  IdResponse,
   DateBasedResponse,
 } from '@app/core';
-import { TokenGenerator } from '../../application/token/token-generator.interface';
-import { Auth, UserIdReq } from '../decorators';
-import { GetUserByIdQuery } from 'apps/api/src/user/application/queries/get-user-by-id';
-import { CreateUserCommand } from 'apps/api/src/user/application/commands/create-user';
-import { AuthResponse } from './responses';
-import { UserResponse } from 'apps/api/src/user/infrastructure/controllers/responses';
-import { MongoUserRepository } from 'apps/api/src/user/infrastructure/repositories';
-import { UserSchema } from 'apps/api/src/user/infrastructure/models/mongo-user.model';
 
 @Controller('auth')
 @ApiTags('Auth')
@@ -56,83 +37,13 @@ export class AuthController {
   constructor(
     @Inject(AUTH_REPOSITORY)
     private readonly repository: CredentialsRepository,
-    @Inject(UUIDGENERATOR)
-    private readonly uuidGenerator: IdGenerator<string>,
     @Inject(BCRYPT_SERVICE)
     private readonly bcryptService: CryptoService,
-    @Inject(JWT_SERVICE)
-    private readonly jwtService: TokenGenerator<string, { id: string }>,
     @Inject(CODE_GENERATOR)
     private readonly codeGenerator: CodeGenerator<string>,
     @Inject(VERIFICATION_EMAIL_HANDLER)
     private readonly verificationEmailHandler: EmailHandler<{ code: string }>,
   ) {}
-
-  @Post('login')
-  @ApiResponse({
-    status: 200,
-    description: 'User logged in successfully',
-    type: AuthResponse,
-  })
-  @ApiResponse({ status: 401, description: 'Unauthorized' })
-  async login(@Body() loginDto: LoginDto) {
-    try {
-      const loginService = new LoginCommand(
-        this.repository,
-        this.jwtService,
-        this.bcryptService,
-      );
-      const loginResult = await loginService.execute(loginDto);
-      const { token, id } = loginResult.unwrap();
-      const service = new GetUserByIdQuery(new MongoUserRepository(new Model(UserSchema)));
-      const result = await service.execute({ id });
-      return { token, user: result.unwrap() };
-    } catch (error) {
-      throw new UnauthorizedException(error.message);
-    }
-  }
-
-  @Post('register')
-  @ApiResponse({
-    status: 201,
-    description: 'User registered successfully',
-    type: IdResponse,
-  })
-  @ApiResponse({ status: 400, description: 'Bad request' })
-  async signUp(@Body() signUpDto: SignUpDto) {
-    try {
-      const signUpService = new SignUpCommand(
-        this.repository,
-        this.jwtService,
-        this.bcryptService,
-      );
-      const service = new CreateUserCommand(
-        new MongoUserRepository(new Model(UserSchema)),
-        this.uuidGenerator,
-      );
-      const result = await service.execute({ ...signUpDto });
-      const id = result.unwrap().id;
-      const signUpResult = await signUpService.execute({ id, ...signUpDto });
-      signUpResult.unwrap();
-      return { id };
-    } catch (error) {
-      throw new UnauthorizedException(error.message);
-    }
-  }
-
-  @Get('current')
-  @Auth()
-  @ApiResponse({
-    status: 200,
-    description: 'User information',
-    type: UserResponse,
-  })
-  @ApiResponse({ status: 401, description: 'Unauthorized' })
-  async currentUser(@UserIdReq() id: string) {
-    const service = new GetUserByIdQuery(new MongoUserRepository(new Model(UserSchema)));
-    const result = await service.execute({ id });
-    return result.unwrap();
-  }
 
   @Post('forget/password')
   @ApiResponse({
