+++ apps/api/src/auth/infrastructure/controllers/dtos/index.ts
@@ -1,5 +1,3 @@
 export { CheckCodeDto } from './check-code.dto';
-export { LoginDto } from './login.dto';
 export { RequestVerificationCodeDto } from './request-code.dto';
 export { ResetPasswordDto } from './reset-password.dto';
-export { SignUpDto } from './register.dto';
