--- apps/api/src/user/infrastructure/controllers/dtos/idex.ts
@@ -1,1 +1,0 @@
-export * from './update-user.dto';
