+++ apps/api/src/user/infrastructure/controllers/dtos/index.ts
@@ -1,0 +1,3 @@
+export { LoginDto } from './login.dto';
+export { SignUpDto } from './register.dto';
+export { UpdateUserDto } from './update-user.dto';
